비밀번호 비교 및 불일치 시 임시 throws Exception 추가

+++ src/main/java/com/example/seatitssu/service/StoreService.java
@@ -79,13 +79,17 @@ public class StoreService {
         return storeInfoResponseDtoList;
     }
 
-    public StoreDeleteResponseDto deleteStore(Long storeId, String password) {
+    public StoreDeleteResponseDto deleteStore(Long storeId, String password) throws Exception {
         Store store = storeRepository.findById(storeId).orElse(null);
 
         if (store == null) {
             System.out.println("=============해당 매장이 존재하지 않습니다.==========");
         }
-        storeRepository.delete(store);
+        if (password.equals(store.getPassword())) {
+            storeRepository.delete(store);
+        } else {
+            throw new Exception("비밀번호가 일치하지 않습니다.");
+        }
         return StoreDeleteResponseDto.of(storeId);
     }
 }
